+++ sdk/spot/websocket/messages/unsubscribe.go
@@ -1,11 +1,13 @@
 package messages
 
-// Request. Subscribe to a topic on a single or multiple currency pairs.
+// Request. Unsubscribe to one or several topics.
 type Unsubscribe struct {
-	// Event type.
+	// Event type. Should be 'unsubscribe'
 	Event string `json:"event"`
 	// Optional - client originated ID reflected in response message.
-	ReqId int `json:"reqid,omitempty"`
+	//
+	// A zero value means no user reference ID will be provided.
+	ReqId int64 `json:"reqid,omitempty"`
 	// Optional - Array of currency pairs. Format of each pair is "A/B", where A and B are
 	// ISO 4217-A3 for standardized assets and popular unique symbol if not standardized.
 	Pairs []string `json:"pair,omitempty"`
@@ -15,15 +17,13 @@ type Unsubscribe struct {
 
 // Subscription details
 type UnsuscribeDetails struct {
-	// Optional - depth associated with book subscription in number of levels each side,
-	// default 10. Cf DepthEnum for values.
+	// Optional - depth associated with book subscription in number of levels each side. Cf. DepthEnum for values.
 	//
-	// A zero value will trigger default behavior.
+	// Default to 10. A zero value will trigger default behavior.
 	Depth int `json:"depth,omitempty"`
-	// Optional - Time interval associated with ohlc subscription in minutes. Default 1.
-	// Cf IntervalEnum for values.
+	// Optional - Time interval associated with ohlc subscription in minutes. Cf. IntervalEnum for values.
 	//
-	// A zero value will trigger default behavior.
+	// Default to 1 minute. A zero value will trigger default behavior.
 	Interval int `json:"interval,omitempty"`
 	// Name of the channel to subscribe to. Cf. ChannelEnum for values.
 	Name string `json:"name"`
